add user profile view rework
